complete the proof of concept

+++ src/component/TagSideBar/TagsWrapper.jsx
@@ -1,23 +1,42 @@
 import { useState, useEffect } from "react";
-import { useSelector } from "react-redux";
+import { useSelector, useDispatch } from "react-redux";
 import { Link } from "react-router-dom";
 import YearTags from "../allTags/YearTags";
 import ResmebleTag from "../allTags/ResmebleTag";
+import { resetUserSelectedImgAction } from "../../action";
 
 const TagsWrapper = () => {
   
   // taking data from the redux store
   const imagesData = useSelector((state) => state.pictureData);
   const tagData = useSelector((state)=> state.tags)
+  const dispatch = useDispatch();
   const [tagObject, setTagObject] = useState({});
+  const finalImgData = useSelector(state => state.userSelectedImg )
   let [imgData, setImgData ] = useState('');
-  
+  const [counter, setCounter] = useState(0);
+
   const uploadHandler = ()=>{
     console.log('upload', imgData);
-    localStorage.clear();    
-    localStorage.setItem('img-details', JSON.stringify(imgData));    
+    localStorage.setItem(counter, JSON.stringify(imgData));
+    setCounter(()=> counter + 1);
+    const data = finalImgData.data;
+    data.forEach( eachElement => {
+      eachElement.remove();
+    });    
+    dispatch(resetUserSelectedImgAction([]));    
   }
-  
+
+  useEffect(()=>{
+    // console.log('final image data to be assigned from tagswrapper : ', finalImgData);
+  }, [finalImgData])
+
+  useEffect(()=>{    
+    if(counter === 0){
+      localStorage.clear();
+    }
+  },[counter])
+   
   useEffect(() => {    
     setTagObject(prevState =>({ ...prevState, ...tagData }));
     setImgData(()=>{
@@ -27,7 +46,6 @@ const TagsWrapper = () => {
         })
       ]
     })
-
     return ()=>{
       setImgData('')
     }
@@ -44,9 +62,9 @@ const TagsWrapper = () => {
       <ResmebleTag/>
     
       <div className="tagsWrapper--buttons">
-        <button className="tagsWrapper--buttons__tagsAssigned-btn">Tags assigned</button>
+        <button className="tagsWrapper--buttons__tagsAssigned-btn" onClick={uploadHandler} >Tags assigned</button>
         <Link to='/filter'>
-          <button className="tagsWrapper--buttons__upload-btn" onClick ={uploadHandler}>Finsih and Upload</button>
+          <button className="tagsWrapper--buttons__upload-btn" >Finsih and Upload</button>
         </Link>
       </div>
